Changed usersController to support all functions neccesary for the routes that have been created for him. UserController now supports get -> that returns all users, and post that adds a new user. Restructered the app, in other words created controllers folder and another folder for routes. In index.php changed the order of initialization and implemented exceptions for errors that could occure while running the server.
Changed usersController to support all functions neccesary for the routes that have been created for him. UserController now supports get -> that returns all users, and post that adds a new user. Restructered the app, in other words created controllers folder and another folder for routes. In index.php changed the order of initialization and implemented exceptions for errors that could occure while running the server.