[Client]
Added sessionStorage to the app so it can cache data in browsers storage, so the user remains logged till he logsout.
Added log out functionality to clear storage
Added login funcitonality to save data in stoarge of the user if he is succesfull logged in.
Added request service, that are wrappers for $http service. (that is a wrapper for xmlHttpReuqest). Every functon from request adds the basic autorization that usses the server.