[Server]
Added baseContorller class that has get, getone, put, post, delete functions implemented so other controllers that iwll inherit (extend, implement) will have by default this funcitons.

Added Exceptions class that will return different exceptions from the server in a format that i want.

Added login functionlaity to the app sever that.

Added a cstatic method Registry that will save the user that makes the request so ever other class will know who accessed this resource.

Added to base model and base controller the functionality to use from registry the data that is saved from the user that is logged in and has autorization.